+++ src/components/functional/unlock-panel/components/index.ts
@@ -1,0 +1,3 @@
+export * from './UnlockPanelFooter';
+export * from './UnlockPanelGroup';
+export * from './UnlockProviderButton';