+++ src/components/functional/unlock-panel/unlock-panel.styles.ts
@@ -1,0 +1,12 @@
+// prettier-ignore
+export default {
+    unlockPanelAnchor: 'unlock-panel-anchor mvx:flex mvx:flex-col mvx:flex-1 mvx:overflow-hidden',
+    unlockPanel: 'unlock-panel mvx:flex mvx:flex-col mvx:h-full mvx:justify-between mvx:overflow-auto mvx:pb-6 mvx:scrollbar-hide',
+    unlockPanelGroups: 'unlock-panel-groups mvx:flex mvx:flex-col mvx:gap-4 mvx:pt-6',
+    unlockButton: 'unlock-button mvx:flex mvx:pl-3 mvx:pr-4 mvx:h-15 mvx:bg-secondary mvx:hover:bg-hover',
+    unlockButtonIcon: 'unlock-button-icon mvx:-order-1 mvx:h-10 mvx:w-10',
+    unlockButtonStatus: 'unlock-button-status mvx:rounded-3xl mvx:py-1 mvx:px-2 mvx:bg-surface',
+    detectedPanelGroup: 'mvx:hidden mvx:sm:flex mvx:sm:flex-col',
+    desktopPanelGroupTitle: 'mvx:hidden mvx:sm:flex',
+    mobilePanelGroupTitle: 'mvx:sm:hidden',
+} satisfies Record<string, string>;